refactor(download): extract Modrinth redirect handling
